pyproject.toml
[project]
name = "moshi"
version = "23.6.6"
description = "Moshi is a spoken language tutor."
readme = "README.md"
requires-python = ">=3.7"
license = {text = "Proprietary"}
keywords = ["language", "ai", "moshi"]
authors = [
  {name = "Eric Kalosa-Kenyon", email = "moshi.feedback@gmail.com"}
]
classifiers = [
  "Development Status :: 2 - Pre-Alpha",
  "Environment :: Web Environment",
  "Programming Language :: Python :: 3",
  "Intended Audience :: Education",
  "Intended Audience :: End Users/Desktop",
  "Topic :: Communications",
  "Topic :: Education",
]
dependencies = [
  "aiortc",
  "aiohttp_session",
  "av",
  "google-auth",
  "google-cloud-translate",
  "google-cloud-texttospeech",
  "jinja2",
  "loguru",
  "numpy",
  "openai",
  "pyfiglet",
]

[project.optional-dependencies]
tests = [
  'pytest',
  # 'pytest-benchmark[histogram]>=3.2.1',
  'pytest-asyncio',
  'scipy',
]
dev = [
  'black',
  'flake8',
  'isort',
]

[project.urls]
Homepage = "https://github.com/ekalosak/moshi/tree/main"
Repository = "https://github.com/ekalosak/moshi.git"

[tool.pytest.ini_options]
markers = [
  "slow: runs more than a second or two to run",
  "openai: access to the _paid_ OpenAI models API",
  "gcloud: access to the _paid_ Google Cloud APIs",
  "asyncio: uses async and await"
]
